+++ index.html
@@ -10,7 +10,7 @@
   <title>Кекстаграм</title>
 </head>
 <script src="js/functions.js"></script>
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 <body>
   <main>
